Remove `strict` field from `SpawnInputExpander` - it is always false.
This field was added in unknown commit (2017) with the goal of prohibiting directory inputs. However, here we are 7 years later and all usages still use `strict = false`. This change removes the dead code for the `true` code path and leaves a `TODO` referencing b/7075837 (deprioritized) in case the effort is ever picked up.

Furthermore, since we don't need to check whether an input is a directory, the `InputMetadataProvider` argument to various methods is no longer necessary.

PiperOrigin-RevId: 599703928
Change-Id: Ibe40a8ae95d1c2ec46207f25d7d331be692eb30a