Add a knob to RunfilesSupplier to override the root of the contained RunfilesTrees.
This seemingly strange contraption is needed because an internal use case does this override. This is fine for now since an Action directly knows about its RunfilesTrees, but the plan is to turn runfiles trees into just Artifacts and plumb the RunfilesTrees in the corresponding RunfilesArtifactValue. This means that such mutation would require flattening the inputs nested set prematurely.

One would think that there is an easier way to do this, however, the decision whether to do this mutation is a function of --test_strategy and needs to be acted upon in the spawn strategy which is a function of --spawn_strategy (et al.) and the only communication conduit between these two is the Spawn (I could have added this method to the Spawn, but I prefer encapsulating it in the RunfilesSupplier so that this change is clearer)

Other alternatives:

1. Signal this in ActionExecutionContext akin to ActionExecutionContext.withOutputsAsInputs(). This would be possible, but would require plumbing the ActionExecutionContext to way more places than it currently is plumbed to.
2. Remove this special-casing. Unfortunately, this is a big incompatible change which I'd like not to have on my critical path.

I considered turning RunfilesSupplier into an abstract class so as to avoid the awkward static method, but it didn't seem to be an obvious win. A default method on the existing interface doesn't work, either, because we use Mockito <2 and that cannot mock interfaces with default methods.

RELNOTES: None.
PiperOrigin-RevId: 599741771
Change-Id: Id8d9baee872e2c0a10bd274e9e43647337eb4591