Implement the focus command.
At a high level, this reduces the amount of retained memory by restricting the build to a predefined/known set of files.

__Details__.

This CL contains the first pass of the GC implementation.

Given a working set of files and target (W, T), `SkyframeFocuser#focus` shrinks the internal `nodeMap` to keep only the nodes that are necessary for the incremental builds of T using changes to W.

__Usage__: To use this, first run a build (e.g. `bazel build //foo`), then run `bazel focus --files=a,b,c`). The `focus` command will use the previous evaluation's top level target (`//foo`) as the root, and `a`, `b`, and `c` as the working set. If the command succeeds, you will see something like: `INFO: Heap: 1419MB -> 764MB (46.18% reduction), Node count: 702292 -> 25458 (96.38% reduction)`.

There are a number of other data structure optimizations as demonstrated in the prototype, missing from this CL. Those will be implemented in later CLs.

PiperOrigin-RevId: 599783801
Change-Id: I9834e1a880bc38136bb98bf8ec18c08761439c1c