Make RunfilesArtifactValue accessible through InputMetadataProvider.
This unfortunately required some re-arrangement of the source tree since now RunfilesArtifactValue is better housed in the .actions package just like FileArtifactValue (or TreeArtifactValue, but I didn't want to spend the energy to move that one)

Currently, runfiles trees can be accessed in two different ways: getInputMetadata() gets the old middleman checksum that changes whenever the runfiles tree does and getRunfilesMetadata() which gets access to the whole RunfilesArtifactValue.

Ideally, we'd make the rule that getInputMetadata() can only be called for non-middleman artifacts, but it has quite a few call sites so I thought it's better not to pollute this already too large change with that and I'm not even sure if we should stop that runfiles trees are regular artifacts anyway.

RELNOTES: None.
PiperOrigin-RevId: 599828274
Change-Id: I6385714f210acc51287b8d9d59620687ad178e5e