site: build.md: Use consistent language
According to https://bazel.build/concepts/build-ref#workspace, the workspace "encompasses the main repo and the set of all defined external repos." Since `//...` excludes targets in external repositories, clarify that it applies to the main repo, not to the workspace.

Similarly, clarify that `//...` and `//:all` refer only to rule targets, not to file targets.

Closes #20963.

PiperOrigin-RevId: 600495266
Change-Id: I36e817990f7041fc15ce2daaac0b628de5254bb5