Remove runfiles trees from action key computation.
This data path has a storied history (main crossroads being unknown commit then unknown commit) but is now unnecessary because for every runfiles tree on the inputs of an action, the corresponding middleman is also there and the checksum of the middleman depends on the contents of the runfiles input manifest (see MiddlemanFactory.createRunfilesMiddleman() and ActionCacheChecker.checkMiddlemanAction()). This means that the exact mapping affects the cache key of the action without having to account for it in computeKey().

There is one aspect of RunfilesTree which is actually removed from the key this way, which is its exact location (it's not the same as that of the middleman artifact), but that is a function of the path to the middleman, so even that works (also, I have no idea how that could possibly happen)

RELNOTES: None.
PiperOrigin-RevId: 600528114
Change-Id: I4523ad14b916a7ebf4f10348b6210564a775e0f1