Compute pretty paths using symlink-creation results instead of disk state
This makes pretty-path printing compatible with
--experimental_convenience_symlinks=log_only, where we'd like to pretend the
symlinks exist and still log relative to them, but not actually create them.

Turns out that instead of inferring symlink state from disk we can pretty easily
pass what we know from symlink creation along to pretty-printing. This also
establishes a happens-before relationship in the code, so the whole "only call
this at the right time" concern should be gone. There is a risk that the on-disk
state could be out of sync with what we think it should be, but I'm having a
hard time thinking of how that would be desired...

PiperOrigin-RevId: 600623440
Change-Id: I9afc9aa7b2e45f911682bde3ce7d20d72a1d9dde