Remove arch attribute from repo markerfile (Related #19563)
Now that we use the marker files for vendoring, each vendored repository has its marker file. To maintain platform independence, we should not include architecture details in these files. So that, if a repository is vendored on one platform and checked in, it should seamlessly work on other platforms, as long as it is up-to-date.

Additionally, this information was initially intended for a specific scenario involving Mac ships. But, we might not need them anymore!

PiperOrigin-RevId: 600743907
Change-Id: Icce5e44d9ca4c9ea16a363268d1e556d19a0397a