Clean up sandboxed workers' directories asynchronously
The setup code now moves the unneeded directory in the workdir out of the way
then deletes it asynchronously. This is faster than waiting for the whole tree
deletion to complete before a worker can be reused.

This cuts down the time spent on workerExecRoot.createFileSystem when building //devtools/buildrank/frontend from an average of 5 seconds to an average of 3 seconds.

RELNOTES:none
PiperOrigin-RevId: 601043337
Change-Id: I47252060fe1a1ea9c9bcfd34fd2987d030f333de