Vendoring (Related #19563)
Vendor command fetches all external repositories into a folder specified by the flag --vendor_dir.
When building using the same flag & value the vendored repos are used as long as they are up to date.

For the initial vendoring or updating existing vendor repositories the new command vendor command is:
`bazel vendor --vendor_dir={vendordir}`

The {vendordir} parameter can be either a full path or just a folder name. Vendored repositories will be located under {workspace}/vendordir.

To perform a build using vendored repositories:
`bazel build {target} --vendor_dir={vendordir}`

In this scenario, the vendor directory will be prioritized for repository retrieval. If the repository is not found or is outdated, the standard process will be followed, which involves checking cached repositories and fetching if not found.

To exclude repositories from the vendoring process:
A file named ".vendorignore" is generated under the vendor directory. By adding the repository name to this file, the vendoring process will ignore it.
Also, any local or configure repos are ignored.

PiperOrigin-RevId: 601099450
Change-Id: I08196c7ad924c1d22918ba09dba2ae774802a691

# Conflicts:
#	src/test/java/com/google/devtools/build/lib/analysis/util/AnalysisTestCase.java
#	src/test/java/com/google/devtools/build/lib/analysis/util/ConfigurationTestCase.java