Implement createReadWriteByteChannel for InMemoryFileSystem.
The upcoming work on disk cache garbage collection will require the ability to read and write to a file while holding a filesystem lock on it. To facilitate testing,
InMemoryFileSystem support must be added. (Work remains to be done to add the locking support.)

Note that InMemoryFileInfo has at least one non-testing use case (in BundledFileSystem) so I put in a bit of effort to minimize buffer reallocations. Performance should be similar to the ByteArrayOutputStream used previously.

Also add some missing test coverage applying to all filesystem implementations.

PiperOrigin-RevId: 601744275
Change-Id: I937f06a8e8d924c19aa681497f12d3d5dabac3af