Remove _build_vars_func from cc_toolchain
The previous implementation used configuration flags from the target where compilation or linking happens. This implementation always uses the flags from the cc_toolchain target.

Because the target (where compilation/linking happens) should more or less directly depend on the cc_toolchain, the values of flags should be the same. Unless somebody puts a configuration transition in the middle.

There's one location I'm aware of, where configuration transition is between the target and cc_toolchain. That is cc_toolchain_forwarder and dummy_configuration_child used in Apple rules. In those instances the desired behaviour is to actually have toolchain variables from the cc_toolchain and not the target.

Testing shows that nothing gets broken and that cc_toolchain variables code may be simplified.

PiperOrigin-RevId: 601751068
Change-Id: I516f5ff27bec14a85dc55a7ab64332c665c6fc65