[7.1.0] Distinguish the disk and remote caches in the action progress status. (#21084)
Previously, they were both displayed as `remote-cache`; there's now a
separate `disk-cache` form. If a combined cache is used, one or both
forms will appear, depending on which caches were looked up.

As a result, the progress status reporting is moved to the individual
cache implementations. While this is kind of unfortunate from an
architectural standpoint, it's likely the best we can do until we recast
cache lookups as spawn strategies (see #19904).

Closes #20935.

PiperOrigin-RevId: 601748051
Change-Id: I03710219973c95d4fca999d931b3513f6d240d94