Evaluate macro implementation functions
https://github.com/bazelbuild/bazel/commit/09eef85a7348d6cf2ddaaad007c1d57ae5a12570 added the ability to declare symbolic macros in packages, but did not run their implementation functions. This CL adds that feature.

For now, implementation functions run synchronously with the call that instantiates the macro. The implementation runs in its own Starlark thread, but it has the side effect of mutating the original Package.Builder, so the targets that it declares are visible to native.existing_rules() in legacy macros. In the future we may make symbolic macro evaluation lazy, in which case the implementation function would run sometime after BUILD evaluation completes, and macro-generated targets would be invisible to native.existing_rules().

MacroClass.java:
- add static helper executeMacroImplementation(). This is analogous to PackageFactory#executeBuildFileImpl(), but I prefer to not weigh that file down with additional code.

SymbolicMacroTest.java:
- tests for signature of implementation function
- test for macro-generated targets. Note that the "my_macro$my_target" naming convention isn't enforced yet.
- test for macro composability
- test for calling existing_rules and glob in a symbolic macro. These will be inverted in a later CL when we ban them, by supplying a different class than Package.Builder to executeMacroImplementation().

Work toward #19922.

PiperOrigin-RevId: 601848482
Change-Id: Ic70fdad9e7d8323c76a3f0125c518547adeee96f