[7.1.0] Fix up permissions error in getInputStream, like we already do for getOutputStream. (#21087)
We're likely reporting permission errors as FileNotFoundException in
some cases, which makes debugging more difficult and might have
correctness implications.

Better yet would be to migrate to the standard
java.nio.file.FileSystemException types, but that's a big and scary
change, and it will have to wait for another day.

PiperOrigin-RevId: 601092545
Change-Id: I1b8777850c946c97e7faf2a47a60255f4e591a3a