Add TargetDefinitionContext
This adds an intermediate class between BazelStarlarkContext and Package.Builder, to prepare the way for a separate builder type for evaluating symbolic macros. A TargetDefinitionContext is anything that can consume the side-effect of defining a target, i.e. either a BUILD file's top-level logic, or a symbolic macro.

NameConflictException is migrated from Package.Builder to TargetDefinitionContext, since in principle it can happen during either BUILD file or symbolic macro evaluation.

Work toward #19922.

PiperOrigin-RevId: 602445896
Change-Id: I6ea14a118480418ae1da3c69973a508261c119f2