Cleanup of ComputedDefault uses and documentation.
Clarify in the docs that ComputedDefault may also read from the package-level defaults of the rule's package and also clarify that ComputedDefault is in-general a deferred evaluation of the default value.

resolvableWithRawAttributes in ComputedDefault can be replaced with a better default check that there are no configured dependencies thus avoiding the need for so many overrides.

PiperOrigin-RevId: 602793682
Change-Id: I3e976a9d817cb3db14f502be70385c20adbef645