Plumb an InputMetadataProvider along with event reporting executed/cached actions and spawns.
This is necessary so that we can remove the call to Action.getRunfilesSupplier() and replace it with iterating over the RunfilesTrees in the input metadata provider.

The actual change to flip the behavior will be done in a followup change so that the functionality change is in a neat small change.

RELNOTES: None.
PiperOrigin-RevId: 603032379
Change-Id: I73f1c76d1ee54b76f7821f29e89acea1dc7ccd6a