[7.1.0] Include the digest hash function in the compact execution log. (#21174)
The digest function is the same for every digest computed by an
invocation. To save space, it is stored once in the Invocation log
entry.

Also embed a Digest proto in File and Spawn messages, with the proviso
that the hash_function_name field will be empty. The size is always
required because the hash alone might not be sufficient to retrieve an
entry from a remote cache.

PiperOrigin-RevId: 602684865
Change-Id: I6f4882051aad9ea72bea0232f8ad4b2cd7b8e335