[7.1.0] Report unresolved symlinks as such in the execution log. (#21177)
In addition, check that an output is of the expected type before
emitting it, and otherwise deem it nonexistent. Strictly speaking, this
is a breaking change for the old (expanded) log formats, but any such
spawn will later cause the respective action to fail, so it doesn't
materially affect the ability to analyze cache misses.

PiperOrigin-RevId: 602724835
Change-Id: Id3ccb2db8362e19d7e26bf8e6bc5cb7fb9478d52