Generate `.ifso` libraries with `llvm-ifs`
RELNOTES: The default Linux C++ toolchain now generates interface
libraries from dynamic libraries to prevent relinking on changes that
only affect the implementation, not the interface of a library. This
requires the `llvm-ifs` tool to be specified via the `BAZEL_LLVM_IFS`
environment variable or available in `PATH`. If this behavior is not
desired, it can be disabled via `--nointerface_shared_objects`.