Use a `BugReporter` in `ActionRewindStrategy`.
In the test case that covers ineffective rewinding (action lost input too many times), use a `RecordingBugReporter` instead of catching the crash thrown by the default bug report in tests. This allows us to test the code path for handling such an exception and assert that the detailed exit code is correct.

PiperOrigin-RevId: 604487369
Change-Id: I39f863b05d27ef7591449814d9243d73db85a668