Account for cache type for uploads in AsyncTaskCache
Currently, the AsyncTaskCache for remote execution uses the digest as a
key to determine whether or not a task has already been run. Normally
this works if there is only a single cache in use. However, when using a
disk cache and a remote cache, some calls try to upload files using only
a single cache through one code path and another cache through another
code path.

This is problematic if only disk caching is used through one code path,
when remote caching is enabled though another code path (ie, uploading
BES referenced artifacts), the files are not uploaded to the remote
cache since the digest is the same. This change uses a String of the
digest and the cache type as the key rather than just the digest to
allow for uploading to both caches.

Update the combined cache test to reflect what should happen:
when remote_build_event_upload=all, *all* files referenced in the BES
stream are uploaded, including build artifacts.

Fixes #20962, #20296

Signed-off-by: Harsh Modi <hmodi@aviatrix.com>