Have `ActionRewindStrategy` throw a more general `ActionRewindException` when an action loses inputs too many times.
`ActionRewindException` is thrown instead of `ActionExecutionException`, which is specific to `ActionExecutionFunction`. When rewinding top-level outputs, we also need to handle the possibility of ineffective rewinding, but an `ActionExecutionException` cannot be constructed since there is no failed `Action` associated with the lost digest. Instead, have the caller catch `ActionRewindException` and construct a relevant exception type.

PiperOrigin-RevId: 604668674
Change-Id: I61a1bfb2f383cebe7852f33b08f0117b3e0d1dd3