Give `ActionRewindStrategy` the responsibility of calling `SkyframeActionExecutor#prepareForRewinding` and emitting `ActionRewoundEvent`.
Moves a bunch of rewinding-specific logic out of `ActionExecutionFunction` and into `ActionRewindStrategy`. My goal is to make `ActionRewindStrategy` more of a "one stop shop" for rewinding logic before adding `CompletionFunction` as another caller.

Also simplifies the return type of `ActionRewindStrategy#prepareRewindPlan` to just `Reset`, since the caller need not be aware of the deps to rewind.

Required breaking out a dedicated `java_library` target for `SkyframeActionExecutor` to avoid a dependency cycle.

PiperOrigin-RevId: 604848689
Change-Id: Ie4c4860d7c9f36fceb904c0e518c907037a392c9