[7.1.0] Avoid exception-based control flow in RAFS#getDigest and RAFS#getFastDigest.
In the same spirit as https://github.com/bazelbuild/bazel/commit/9a86600f68dfe1d802d4e7b2a757e69f20a308d0, avoid unnecessarily allocating a FileNotFoundException when it's not going to be propagated. It makes getFastDigest rather slow in the failure case.

This requires changing the contract of statUnchecked (now statInternal) to return null on not found, leaving the decision of whether to throw an exception up to the caller. I/O exceptions other than not found are always propagated, as before.

Also improve consistency in terminology:
  * "deferencing a path" => "canonicalizing a path"
  * "in-memory filesystem" => "remote output tree"

PiperOrigin-RevId: 604973882
Change-Id: I1f78b8c5cce3707eb0608a8c0c85a2b04445a510