Refactor CgroupsInfo to decouple the implementations between v1 and v2.
Also include an InvalidCgroupsInfo implementation that allows us to represent a cgroup that was attempted to be created but not successful (rather than using null to represent it).

The general way for clients to use this is to create the CgroupsInfo with the methods `CgroupsInfo.getBlazeSpawnsCgroup()`, `CgroupsInfo.getBlazeSpawnsCgroup().createIndividualSpawnCgroup()`. Then check the returned cgroup instance with `exists()` to ensure that it is created (and not invalid) before continuing with other actions e.g. `addProcess()`.

PiperOrigin-RevId: 605269206
Change-Id: I69fe32e3087e37e0705b840e10f07f37160402d0