Add more Pressure Stall Indicator(PSI) metrics to Blazel trace profiler.
There are 2 types of PSI metrics: full and some. The "some" line indicates the share of time in which at least some tasks are stalled on a given resource. The "full" line indicates the share of time in which all non-idle tasks are stalled on a given resource simultaneously.There is no full PSI metric for CPU resource.

In this change I've added CPU PSI metric. Also I've added stacked memory and IO metrics with split on some/full types.

PiperOrigin-RevId: 605569491
Change-Id: Ia37faa60e783d8faf8e2bd614cf2899b9704cb5f