Maintain a separate lockfile for HEAD Bazel
The MODULE.bazel.lock file we have in our source tree is generated by the version of Bazel specified in the .bazelversion file. We currently use this lockfile for our distribution archives, which means whenever we increment the lockfile version, we need to manually transform the checked-in lockfile to make it work with HEAD Bazel too.

This commit just checks in a lockfile for HEAD Bazel and removes that hack. This separate lockfile is updated by running `bazel run //src/test/tools/bzlmod:update_default_lock_file`. If it goes out of date, the distfile tests will complain.