Only register `.jdeps` cache `ActionContext` when used
By registering `JavaCompileActionContext` (renamed to
`JdepsCachingActionContext` in this change to better reflect its only
purpose) only when using `--java_classpath=bazel`, the
`insertDependencies` field of `JavaHeaderCompileAction` becomes
obsolete. Combined with the existing 3 bytes of padding, this makes room
for a new 4-byte field that will be used in a rollforward of
fd196bf03f4301cc3fbfc307570637f4de71c82c.