[7.1.0] Fix a hanging issue with skymeld & `--combined_report=lcov`. (#21271)
With Skymeld, the combined report is done right before the end of the
build. This works fine except for a buggy case:
- `--nokeep_going`
- when there's an error that the Coverage action transitively depends on
- The error would try to interrupt the build, and since we evaluate
coverage actions uninterruptibly (consistent with the noskymeld
behavior), we're stuck in an infinite loop [1].

This happened because we did not fail fast like we should have. This CL
fixes the issue by checking, before generating the combined report,
whether we should fail fast.

Fixes https://github.com/bazelbuild/bazel/issues/21154

[1]
https://github.com/bazelbuild/bazel/blob/026f493a5a403fa5d770cae0b3a3baf8dcf33488/src/main/java/com/google/devtools/build/lib/concurrent/Uninterruptibles.java#L33-L39

Commit
https://github.com/bazelbuild/bazel/commit/37247d53dc828ad9d03cf4a1537d1db5d83caf99

PiperOrigin-RevId: 605572100
Change-Id: I8d57dacca58358771799161352819ec65bef6ac2

Co-authored-by: Googler <leba@google.com>