Generate a lockfile for the distribution archive on the fly (third_party changes from PR #21302)
The MODULE.bazel.lock file we have in our source tree is generated by the version of Bazel specified in the .bazelversion file. We currently use this lockfile for our distribution archives, which means whenever we increment the lockfile version, we need to manually transform the checked-in lockfile to make it work with HEAD Bazel too.

This commit generates a lockfile for HEAD Bazel on the fly and uses that for the distribution archive. This lockfile doesn't have any module extension content, since we only need to make sure that no network access happens during tests that use the distribution archive. See https://github.com/bazelbuild/bazel/pull/21283#issuecomment-1938398252 for more context.

Change-Id: I5266bd1caf2e4cc3d7d18cb8695af73488064ae1
Signed-off-by: Xudong Yang <wyv@google.com>