Handle lost outputs in partially built top-level targets.
A partially built top-level target can occur if some but not all of its artifacts are successfully built. When this happens, the built artifacts are still reported in the BEP, so we need to treat them the same as outputs from a fully successful target.

Note that this change only handles the `--keep_going` case properly. Attempting to initiate rewinding during error bubbling won't work, so we'll need a different approach for `--nokeep_going`, which I will implement separately.

In addition to the new test case for this scenario, I added assertions to existing top-level output rewinding test cases to verify artifact reporting.

PiperOrigin-RevId: 606391715
Change-Id: Iaac24df75bfb15f10c046704dfb7ff1d807c1b8f