Make it possible to override the path of runfiles trees.
This is needed for a quirk in the remote execution implementation at Google. Previously, its implementation was in Google-internal code but removing RunfilesSupplier requires that we move this code here.

It's quite a bit of of boilerplate, but in exchange we can delete RunfilesSupplier.getRunfilesDirOverride() and remove the "possibly incorrect" path from getPossiblyIncorrectExecPath(). This also encapsulates knowledge about the runfiles trees in InputMetadataProvider, where it belongs.

RELNOTES: None.
PiperOrigin-RevId: 606613420
Change-Id: I4d9220ba02758a100f7115b2f3a5995c4c3e4ac4