[7.1.0] Optimize RemoteActionFileSystem#resolveSymbolicLinks by caching intermediate results in a trie. (#21333)
When scanning a filesystem tree, resolveSymbolicLinks does O(M*N) work,
where M is the number of components in a file path and N is the number
of files. This CL makes it O(M+N) instead. This makes large output tree
artifacts (~250k files) much more efficient to scan (from ~45s to ~9s in
a particular example; additional optimizations are possible and will be
made in a followup CL).

Related to https://github.com/bazelbuild/bazel/issues/17009.

PiperOrigin-RevId: 606259673
Change-Id: Icf781a78b3271196e0029e3049d969a9e6073906