Add behavior to `CompletionContext` to expand artifacts and calculate owners.
This will be used for rewinding special top-level artifacts (tree artifacts and filesets).

Care was taken to share as much code as possible between the existing `CompletionContext#visitArtifacts` which supports BEP infrastructure (visitor pattern) and the new methods to service `ImportantOutputHandler` (takes a `Collection`). This way we match BEP reporting semantics.

PiperOrigin-RevId: 606800080
Change-Id: Ifba83cc77d930367ad585b51b8b564d41385cfd3