Handle lost outputs for partially built targets in `--nokeep_going` mode.
Ensure that an event is always posted prior to attempting rewinding for partially built targets. This works because completion events always return true for `storeForReplay()`, which means they won't be emitted until skyframe can guarantee that the function won't be attempted again on this evaluation (either all deps are complete or we're in error bubbling). Thus, only the last posted event is emitted for a given `SkyKey`. Note that BEP handling of partially built targets (unknown commit) already relies on this behavior.

To ensure that this approach works, document and check that `Completor` only returns events with `storeForReplay() = true`, and check for duplicate events in rewinding tests.

PiperOrigin-RevId: 607134767
Change-Id: Id1b18a16bc172e58f591f76e746b516b76289adc