[7.1.0] Collect directory contents in parallel in CompactSpawnLogContext. (#21361)
Significantly reduces time spent computing a log entry for large tree
artifacts (in a particular execution with ~250k files, the time is
reduced from ~72s to ~6s).

This is the same idea as
https://github.com/bazelbuild/bazel/commit/368bf117513731909337b6e2bd5082c046399506,
later improved by
https://github.com/bazelbuild/bazel/commit/df07d275ce17f14224af0f1a5862755a7db44397.

In theory, the optimization could also be applied to
ExpandedSpawnLogContext. However, I don't want to invest resources into
improving the old log formats at this time.

Also sort the directory contents alphabetically, fixing an oversight in
the previous implementation.

PiperOrigin-RevId: 607126723
Change-Id: I0ebcdbd5e897e25b5b0c28a642e50acae2fd6730