Handle symlinks in a more consistent manner in UploadManifest.
This CL makes the following behavioral changes:

1. `allowDanglingSymlinks` now also applies to symlinks inside an output directory. That is, when `allowDanglingSymlinks` is false, a dangling symlink inside an output directory will cause an error, matching the preexisting behavior for a dangling symlink outside of an output directory.

2. The handling of a dangling symlink when `followSymlinks` is true is now consistent inside and outside of an output directory: it gets uploaded as a symlink nevertheless, as long as `allowDanglingSymlinks` is true (previously, it was always an error inside of an output directory).

These changes are extremely unlikely to break anyone: (1) only affects the pre-flip state of an incompatible flag (`--incompatible_remote_dangling_symlinks`) we already plan to delete, and (2) turns an upload error into a non-error (and, in addition, the action owning the spawn would fail anyway, because a tree artifact must not contain dangling symlinks).

A third inconsistency is noted in the now amended flag description for `--incompatible_remote_symlinks`, but won't be fixed in this CL, because it's riskier and will likely require yet another incompatible flag flip.

Also take the opportunity to rewrite the logic in a more readable, less branchy way.

PiperOrigin-RevId: 607316436
Change-Id: Ice7c04601e8594241a330670660d4e0791268cb2