Handle expanded artifacts (tree artifacts and filesets) in top-level output rewinding.
`ActionRewindStrategy` already has handling for these types of artifacts, provided that it is passed an `ActionInputDepOwners` that tracks "owners" mappings for all such lost artifacts. The mapping represents the relationship between an expanded `ActionInput` and its owning `Artifact` (either a tree or a fileset).

In order to match BEP reporting semantics, add behavior to `CompletionContext` to expand artifacts and calculate owners. Care was taken to share as much code as possible between the existing `CompletionContext#visitArtifacts` which supports BEP infrastructure (visitor pattern) and `ImportantOutputHandler` (takes a `Collection`).

PiperOrigin-RevId: 607379568
Change-Id: I4199982dc65342b9af45cf4d57206442bd6307be