Add a new `skyframeGlobsExecutorService` ForkJoinPool to `AnalysisExecutor`
In this change, I intentionally set default flag value (aka thread number) in `PackageOptions` to be 0, so this change will be a no-op for now.

In the future, this FJP will be injected into `GlobsFunction` so that in-parallel `StateMachine`s are pushed to be run on this FJP concurrently. A positive default flag value (thread count) should be determined at that time.

I also have considered using an existing FJP (`cpuHeavyPoolExecutorService`) instead of creating a new one. And eventually chose to create a new one. There are several reasons for this decision:

1. A new FJP with a limited number of threads only adds a very small resource usage to blaze, which is acceptable.
2. Reusing `cpuHeavyPoolExecutorService` requires heavy renaming (especially we will need to change the flag name), which should be avoided.
3. If some logic ends of blocking the threads in the same thread pool I try to add threads to, a deadlock can happen.

PiperOrigin-RevId: 607732468
Change-Id: Ib6226a545f7efdc32c3213cc74425cca75f3e508